# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SRM Credit Engine",
    "version": "0.5.0",
    "description": "Cadastro, prévia CSV, consulta, simulação e liquidação individual com idempotência. Valores decimais em strings. Data de negócio America/Sao_Paulo. Ambiente demonstrativo local sem autenticação; não expor publicamente."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/settlements": {
      "post": {
        "summary": "Liquidar integralmente com idempotência e reconfirmação",
        "description": "Primeira execução 201; repetição semanticamente igual da chave/pedido concluído 200, antes de recalcular/validar câmbio. Mesma chave com pedido diferente: 409 IDEMPOTENCY_CONFLICT. Outra chave para SETTLED: 409 ALREADY_SETTLED. Concorrência em andamento: 409 OPERATION_IN_PROGRESS; não mudar chave/payload ao repetir. Divergência financeira: 409 CONDITIONS_CHANGED com current; confirmar novamente é uma nova intenção/chave. Chaves de operações concluídas ficam retidas sem TTL. Nenhuma transferência bancária real.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID gerado pelo cliente uma vez por intenção; reutilizar em resultado incerto."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replay da liquidação original",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settlement"
                }
              }
            }
          },
          "201": {
            "description": "Liquidação criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settlement"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URI relativa para consulta"
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recebível não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito; CONDITIONS_CHANGED inclui current. OPERATION_IN_PROGRESS preserva a intenção para retry manual.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/ConditionsError"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Cotação ausente/expirada impede liquidação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Falha técnica; não presumir resultado pelo status de rede",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Consultar snapshots de liquidações por settledAt e id decrescentes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Índice base zero. page × size deve ser <= 2147483647."
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Início inclusivo em settledAt. ISO 8601 com horário e offset; anos 0001–9999, precisão máxima de microssegundos.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Fim exclusivo em settledAt. Mesma precisão de from; deve ser posterior ao início quando ambos são informados.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "assignorId",
            "in": "query",
            "required": false,
            "description": "ID do cedente registrado no snapshot.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "description": "Moeda de pagamento do snapshot.",
            "schema": {
              "type": "string",
              "enum": [
                "BRL",
                "USD"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementPage"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro técnico; sem dados internos na resposta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Filtros opcionais combináveis. Não recalcula históricos. Cedente inexistente ou página sem resultados retorna 200 com content vazio. Paginação por offset não congela a coleção entre requisições."
      }
    },
    "/settlements/{id}": {
      "get": {
        "summary": "Consultar snapshot imutável de liquidação",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot persistido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settlement"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Liquidação não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Falha técnica",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receivables/{id}/simulations": {
      "post": {
        "summary": "Simular título PENDING e obter expectedConditions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Título, cálculo e condições para confirmação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivableSimulation"
                }
              }
            }
          },
          "400": {
            "description": "ID, vencimento ou prazo inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Título não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Título já liquidado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Cotação indisponível/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Falha técnica",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ReceivableType": {
        "type": "string",
        "enum": [
          "DUPLICATA_MERCANTIL",
          "CHEQUE_PRE_DATADO"
        ]
      },
      "PaymentCurrency": {
        "type": "string",
        "enum": [
          "BRL",
          "USD"
        ]
      },
      "Status": {
        "type": "string",
        "enum": [
          "PENDING",
          "SETTLED"
        ]
      },
...
      "Receivable": {
        "type": "object",
        "required": [
          "id",
          "assignorId",
          "titleCode",
          "type",
          "faceValue",
          "dueDate",
          "paymentCurrency",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "assignorId": {
            "type": "string",
            "format": "uuid"
          },
          "titleCode": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ReceivableType"
          },
          "faceValue": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]{2}$",
            "example": "92859.94"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "paymentCurrency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Simulation": {
        "type": "object",
        "required": [
          "type",
          "referenceDate",
          "dueDate",
          "calculatedAt",
          "faceValueBrl",
          "termMonths",
          "baseRate",
          "spread",
          "effectiveRate",
          "presentValueBrl",
          "discountBrl",
          "paymentCurrency",
          "exchangeRateUsed",
          "finalAmount",
          "exchangeRateId",
          "exchangeRateValidFrom"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ReceivableType"
          },
          "referenceDate": {
            "type": "string",
            "format": "date"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "calculatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "faceValueBrl": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]{2}$",
            "example": "92859.94"
          },
          "termMonths": {
            "type": "integer",
            "minimum": 0,
            "maximum": 120
          },
          "baseRate": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "Decimal em texto, ponto decimal, sem expoente; rejeita dígitos significativos além da precisão/escala."
          },
          "spread": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "Decimal em texto, ponto decimal, sem expoente; rejeita dígitos significativos além da precisão/escala."
          },
          "effectiveRate": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "Decimal em texto, ponto decimal, sem expoente; rejeita dígitos significativos além da precisão/escala."
          },
          "presentValueBrl": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]{2}$",
            "example": "92859.94"
          },
          "discountBrl": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]{2}$",
            "example": "92859.94"
          },
          "paymentCurrency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "exchangeRateUsed": {
            "type": [
              "string",
              "null"
            ]
          },
          "finalAmount": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]{2}$",
            "example": "92859.94"
          },
          "exchangeRateId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "exchangeRateValidFrom": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "fieldErrors"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ExpectedConditions": {
        "type": "object",
        "required": [
          "assignorId",
          "titleCode",
          "type",
          "dueDate",
          "termMonths",
          "paymentCurrency",
          "faceValueBrl",
          "effectiveRate",
          "presentValueBrl",
          "discountBrl",
          "finalAmount"
        ],
        "properties": {
          "assignorId": {
            "type": "string",
            "format": "uuid"
          },
          "titleCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "$ref": "#/components/schemas/ReceivableType"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "termMonths": {
            "type": "integer",
            "minimum": 0,
            "maximum": 120
          },
          "paymentCurrency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "faceValueBrl": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "NUMERIC(19,2); normaliza zeros equivalentes sem arredondamento."
          },
          "effectiveRate": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "Decimal em texto, ponto decimal, sem expoente; rejeita dígitos significativos além da precisão/escala."
          },
          "presentValueBrl": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "NUMERIC(19,2); normaliza zeros equivalentes sem arredondamento."
          },
          "discountBrl": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "NUMERIC(19,2); normaliza zeros equivalentes sem arredondamento."
          },
          "finalAmount": {
            "type": "string",
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "NUMERIC(19,2); normaliza zeros equivalentes sem arredondamento."
          },
          "exchangeRateUsed": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "maxLength": 64,
            "description": "USD exige cotação positiva NUMERIC(24,10); BRL exige null ou campo omitido."
          }
        },
        "description": "Condições exibidas ao usuário. Decimais são normalizados à escala do banco sem arredondar; escalas equivalentes têm a mesma intenção. Não contém IDs/vigência de cotação ou timestamps. Valores enviados não determinam o pagamento."
      },
      "SettlementRequest": {
        "type": "object",
        "required": [
          "receivableId",
          "expectedConditions"
        ],
        "properties": {
          "receivableId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedConditions": {
            "$ref": "#/components/schemas/ExpectedConditions"
          }
        }
      },
      "ReceivableSimulation": {
        "type": "object",
        "required": [
          "receivable",
          "simulation",
          "expectedConditions"
        ],
        "properties": {
          "receivable": {
            "$ref": "#/components/schemas/Receivable"
          },
          "simulation": {
            "$ref": "#/components/schemas/Simulation"
          },
          "expectedConditions": {
            "$ref": "#/components/schemas/ExpectedConditions"
          }
        }
      },
      "Settlement": {
        "type": "object",
        "required": [
          "id",
          "receivableId",
          "idempotencyKey",
          "assignorId",
          "assignorName",
          "titleCode",
          "settledAt",
          "calculation"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "receivableId": {
            "type": "string",
            "format": "uuid"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "assignorId": {
            "type": "string",
            "format": "uuid"
          },
          "assignorName": {
            "type": "string"
          },
          "titleCode": {
            "type": "string"
          },
          "settledAt": {
            "type": "string",
            "format": "date-time"
          },
          "calculation": {
            "$ref": "#/components/schemas/Simulation"
          }
        }
      },
      "ConditionsError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "fieldErrors",
          "current"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "current": {
            "$ref": "#/components/schemas/ReceivableSimulation"
          }
        }
      },
      "SettlementPage": {
        "type": "object",
        "required": [
          "content",
          "page",
          "size",
          "totalElements",
          "totalPages"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Settlement"
            }
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      }
    }
  }
}